core/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.3
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-etc==1.4.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
phonenumbers==8.13.17
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
